requirements.txt
openai>=1.59.0,<2
pandas>=2.0.0,<3
openpyxl>=3.1.0,<4
flask>=3.0.0,<4
pyyaml>=6.0.0,<7
python-dateutil>=2.9.0,<3
jinja2>=3.1.0,<4
pytest>=7.4.0,<8
